+++ src/cli/power.rs
@@ -2,9 +2,9 @@ use clap::Subcommand;
 use pk_cli_core::CliError;
 use serde_json::json;
 
-use super::output::emit_one;
 use super::Ctx;
 use crate::resolve;
+use pk_cli_core::output::emit_one;
 
 #[derive(Subcommand, Debug)]
 pub enum PowerCommand {